refactor(all): relocate files, remove redundant files.
