+++ src/App.tsx
@@ -2,7 +2,7 @@ import { Reset } from 'styled-reset';
 import { createGlobalStyle } from 'styled-components';
 import { BrowserRouter, Routes, Route, Navigate } from 'react-router-dom';
 import React from 'react';
-import { GameMap } from './pages/gameMap/GameMap';
+import { Main } from './pages/main/Main';
 import AuthRoute from './component/AuthRoute';
 import Header from './component/Header';
 import { Landing } from './pages/landing/Landing';
@@ -17,7 +17,7 @@ const App: React.FunctionComponent<IApplicationProps> = (props) => {
       <AuthRoute>
         <Header />
         <Routes>
-          <Route path="/city" element={<GameMap />} />
+          <Route path="/city" element={<Main />} />
           <Route path="/landing" element={<Landing />} />
           <Route path="*" element={<Navigate to="/landing" replace />} />
         </Routes>
