+++ src/pages/city/housesSvg/HouseGrid.tsx
