+++ src/pages/city/housesSvg/HouseOfClothes.tsx
