+++ src/pages/city/housesSvg/HouseOfDrinks.tsx
