+++ src/pages/city/housesSvg/HouseOfPlants.tsx
