+++ src/redux/api/ledgerListAPI.ts
@@ -1,4 +1,4 @@
-import { db } from '../../config/firebase';
+import { db } from '../../utils/firebase';
 import {
   getDocs,
   collection,
